feat: restructure project and use cmake

+++ include/ga/parameters.h
@@ -1,11 +1,11 @@
 #pragma once
 
 extern const int SEED;
-extern const int POPULATION_SIZE;
-extern const int TOURNAMENT_SIZE;
+extern const size_t POPULATION_SIZE;
+extern const size_t TOURNAMENT_SIZE;
 extern const float MUTATION_RATE;
 extern const float LOCAL_OPT_RATE;
-extern const int ELITE_SIZE;
+extern const size_t ELITE_SIZE;
 extern const int MAX_GENERATIONS;
 extern const int MAX_THREADS;
